requirements.txt
numpy>=2.3.3
scipy>=1.15.3
pynrrd>=1.1.3
torch>=2.6.0
matplotlib>=3.10.5